now it boots on windows again

+++ scripts/buildpackage.py
@@ -13,8 +13,8 @@ from deb import make_deb
 from shutil import copyfile
 from shutil import rmtree
 
-def windows_install(d):
-	if d.yesno("make install?") == d.OK:
+def build_windows_install(d):
+	if d.yesno("Build files for windows installer?") == d.OK:
 		dll_opengl_path="/home/rod/windll/opengl_dlls/"
 		dll_compiled_path="/home/rod/windll/compiled_dlls/"
 
@@ -33,15 +33,23 @@ def windows_install(d):
 		for file in os.listdir(dll_compiled_path):
 			copyfile(os.path.join(dll_compiled_path,file), os.path.join(output_path,"gpvdm",file))
 
-		publish_dir="/home/rod/windows/share/pub"
-		if os.path.isdir(publish_dir)==True:
-			rmtree(publish_dir)
+		windows_share_path="/home/rod/windows/share"
+		if os.path.isdir(windows_share_path)==True:
+			ret=d.inputbox("Copy files to "+windows_share_path+"?", init=windows_share_path)
+			if ret[0]=="ok":
+				windows_share_path=ret[1]
+				print
+				#yesno("Copy files to "+windows_share_path+"?") == d.OK:
+				publish_dir=os.path.join(windows_share_path,"pub")
+				if os.path.isdir(publish_dir)==True:
+					rmtree(publish_dir)
 
-		#os.mkdir(publish_dir)
-		os.remove("./pub/gpvdm/gpvdm")
-		shutil.copytree("./pub/gpvdm", publish_dir, symlinks=False)
+				#os.mkdir(publish_dir)
+				os.remove("./pub/gpvdm/gpvdm")
+				shutil.copytree("./pub/gpvdm", publish_dir, symlinks=False)
+				d.msgbox("The build has been placed in: "+windows_share_path)
 
-def distributable(d):
+def buildpackage_menu(d):
 	if os.geteuid() == 0:
 		d.msgbox("Don't run me as root.")
 		return
@@ -63,7 +71,7 @@ def distributable(d):
 				make_deb(d)
 
 			if tag=="(win)":
-				windows_install(d)
+				build_windows_install(d)
 
 		else:
 			return
